agent-driven: Move validate_property_name to ucd_helper.rs
